streamlit/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=streamlit/requirements.txt streamlit/requirements.in
#

altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.4.0
    # via streamlit
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.7
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via python-aapy
numpy==2.0.1
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   altair
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pillow==10.4.0
    # via streamlit
protobuf==5.27.2
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.8.2
    # via python-aapy
pydantic-core==2.20.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-aapy==2.0.3
    # via -r streamlit/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via python-aapy
pytz==2024.1
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   python-aapy
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via python-aapy
rich==13.7.1
    # via streamlit
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.37.0
    # via streamlit-tags
streamlit-tags==1.2.8
    # via -r streamlit/requirements.in
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via python-aapy
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via streamlit
